done: leanaide_server and streamlit works
