+++ leanaide_server.py
@@ -1,118 +1,74 @@
-# Code from Gemini 2.0 Flash Experimental
-#
-import http.server
-import subprocess
-import json
+import multiprocessing
 import os
-import threading
-import queue
+import signal
+import socket
+import subprocess
 import sys
-PORT = int(os.environ.get("LEANAIDE_PORT", 7654))
-HOST = os.environ.get("HOST", "localhost")  
-COMMAND = os.environ.get("LEANAIDE_COMMAND", "lake exe leanaide_process")
-for arg in sys.argv[1:]:
-    COMMAND += " " + arg
+import time
+from pathlib import Path
 
-print(f"Command: {COMMAND}")
-process = None
-process_lock = threading.Lock()
-output_queue = queue.Queue()
+STREAMLIT_PORT = 8501
+LEANAIDE_PORT = int(os.environ.get("LEANAIDE_PORT", 7654))
 
-def process_reader(process, output_queue):
-    while True:
-        line = process.stdout.readline()
-        if not line:
-            break  # Process terminated
-        output_queue.put(line.strip())
+home_dir = str(Path(__file__).resolve().parent)
+serv_dir = os.path.join(home_dir, "server")
 
-def process_error_reader(process):  # New function for stderr
-    while True:
-        line = process.stderr.readline()
-        if not line:
-            break  # Process terminated
-        print(f"Process stderr: {line.strip()}", file=sys.stderr)  # Print to server's stderr
+STREAMLIT_FILE = os.path.join(serv_dir, "streamlit_ui.py")
+SERVER_FILE = os.path.join(serv_dir, "api_server.py")
 
-class Handler(http.server.BaseHTTPRequestHandler):
-    def do_POST(self):
-        global process, process_lock
-        content_length = int(self.headers['Content-Length'])
-        post_data = self.rfile.read(content_length).decode('utf-8')
+def is_port_in_use(port: int) -> bool:
+    """Check if a port is already in use"""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        return s.connect_ex(('localhost', port)) == 0
 
-        try:
-            with process_lock:
-                if process is None or process.poll() is not None:
-                    try:
-                        process = subprocess.Popen(
-                            COMMAND.split(),
-                            stdin=subprocess.PIPE,
-                            stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE,
-                            text=True,
-                            bufsize=1 # Line buffering
-                        )
-                        threading.Thread(target=process_reader, args=(process, output_queue), daemon=True).start()
-                        threading.Thread(target=process_error_reader, args=(process,), daemon=True).start()  # Start stderr thread
-                    except FileNotFoundError:
-                        self.send_response(500)
-                        self.send_header('Content-type', 'text/plain')
-                        self.end_headers()
-                        self.wfile.write(f"Command not found: {COMMAND}".encode())
-                        return
-            try:
-                json_data = json.loads(post_data)
-                process.stdin.write(json.dumps(json_data) + '\n')
-                process.stdin.flush()
+def run_server_api():
+    """Run Server server using uvicorn"""
+    subprocess.run([sys.executable, SERVER_FILE])
 
-                try:
-                    output = output_queue.get(timeout=6000)  # Timeout after 6000 seconds
-                    self.send_response(200)
-                    self.send_header('Content-type', 'application/json')
-                    self.end_headers()
-                    self.wfile.write(output.encode())
-                except queue.Empty:
-                    self.send_response(504)  # Gateway Timeout
-                    self.send_header('Content-type', 'text/plain')
-                    self.end_headers()
-                    self.wfile.write("Process timed out".encode())
-            except json.JSONDecodeError:
-                self.send_response(400)
-                self.send_header('Content-type', 'text/plain')
-                self.end_headers()
-                self.wfile.write("Invalid JSON".encode())
-            except BrokenPipeError:
-                with process_lock:
-                    process = None
-                self.send_response(500)
-                self.send_header('Content-type', 'text/plain')
-                self.end_headers()
-                self.wfile.write("Process crashed".encode())
-        except Exception as e:
-            self.send_response(500)
-            self.send_header('Content-type', 'text/plain')
-            self.end_headers()
-            self.wfile.write(str(e).encode())
+def run_streamlit():
+    """Run Streamlit app on port STREAMLIT_PORT only"""
+    # STREAMLIT_PORT is used by our main Streamlit, this prevents another instance of same streamlit being made
+    if is_port_in_use(STREAMLIT_PORT):
+        # print("❌ Port STREAMLIT_PORT is already in use! Close other Streamlit instances first.")
+        return
 
-    def do_GET(self):
-        self.send_response(200)
-        self.send_header('Content-type', 'text/plain')
-        self.end_headers()
-        self.wfile.write("Send a POST request with JSON to interact with the process.".encode())
+    # Force Streamlit to use only port STREAMLIT_PORT
+    subprocess.run([
+        sys.executable, "-m",
+        "streamlit", "run", STREAMLIT_FILE,
+        f"--server.port={STREAMLIT_PORT}",
+        "--server.headless=true",
+        "--browser.gatherUsageStats=false"
+    ])
 
-
-def run(server_class=http.server.HTTPServer, handler_class=Handler, port=PORT, host=HOST):
-    server_address = (host, port) # Use the host
-    httpd = server_class(server_address, handler_class)
-    print(f"Starting httpd on port {port}, command: {COMMAND}")
-    try:
-        httpd.serve_forever()
-    except KeyboardInterrupt:
-        print("Stopping server...")
-    finally:
-        global process
-        if process:
-            process.terminate()
-            process.wait()
-            print("Process terminated.")
+def signal_handler(sig, frame):
+    """Handle Ctrl+C to terminate both processes"""
+    print("\nShutting down servers...")
+    sys.exit(0)
 
 if __name__ == "__main__":
-    run(host=HOST)
+    # Set up signal handler
+    signal.signal(signal.SIGINT, signal_handler)
+
+    print("\n\033[1;34mStarting servers (single-port mode)\033[0m\n")
+    print(f"\033[1;34mAPI Server:\033[0m http://{os.environ.get("HOST", "localhost")}:{LEANAIDE_PORT}")
+    print()
+
+    # Start processes
+    serv_process = multiprocessing.Process(target=run_server_api)
+    streamlit_process = multiprocessing.Process(target=run_streamlit)
+
+    serv_process.start()
+    time.sleep(0.5)  # Brief delay to ensure server starts first
+    streamlit_process.start()
+
+    try:
+        # Keep main process alive
+        while True:
+            pass
+    except KeyboardInterrupt:
+        print("\nShutting down...")
+        serv_process.terminate()
+        streamlit_process.terminate()
+        serv_process.join()
+        streamlit_process.join()
