+++ server/pyproject.toml
@@ -5,10 +5,8 @@ description = "Leanaide Streamlit Server"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-    "fastapi>=0.115.12",
     "pyperclip>=1.9.0",
     "python-dotenv>=1.1.0",
     "streamlit>=1.45.1",
-    "uvicorn>=0.34.3",
     "watchdog>=6.0.0",
 ]
