+++ server/requirements.txt
@@ -1,7 +1,5 @@
 # Requirements for the project
-fastapi>=0.115.12
 pyperclip>=1.9.0
 python-dotenv>=1.1.0
 streamlit>=1.45.1
 watchdog>=6.0.0
-uvicorn>=0.34.3
